+++ server/src/domain/post/post.controller.ts
@@ -71,20 +71,4 @@ export class PostController {
       message: '글 작성에 성공했습니다.',
     };
   }
-
-  @Post(':postId/likes')
-  @UseGuards(AuthGuard('jwt'))
-  @HttpCode(HttpStatus.CREATED)
-  async likes(@Req() req: Request, @Param('postId') postId: number) {
-    const userId = req.user['id'];
-    await this.postService.addLikes(userId, postId);
-  }
-
-  @Delete(':postId/likes')
-  @UseGuards(AuthGuard('jwt'))
-  @HttpCode(HttpStatus.NO_CONTENT)
-  async cancelLikes(@Req() req: Request, @Param('postId') postId: number) {
-    const userId = req.user['id'];
-    await this.postService.cancelLikes(userId, postId);
-  }
 }
