+++ server/src/domain/post/post.repository.ts
@@ -9,6 +9,7 @@ export class PostRepository extends Repository<Post> {
     super(Post, dataSource.createEntityManager());
   }
 
+  // TODO 좋아요 눌렀는지 보내주기
   async findByIdUsingCondition(
     lastId: number,
     postIdsFiltered: number[],
